IDEA-59587 "Git: after file is moved its status is 'added' until refresh". Using custom isUnder().
When we move a file we get a VcsDirtyScope with old and new FilePaths, but unfortunately the virtual file in the FilePath is refreshed and points to the new position which makes FilePathImpl#isUnder useless in our case. So we are using custom isUnder() which considers only IOFiles (which are not updated).