IDEA-67875 Fix update and push for empty repositories.
When the remote doesn't have the branch which current branch tracks (for example, an empty repository was cloned, nothing is on origin/master yet), we shouldn't query this remote branch.

GitBranch.exists() checks if the branch really exists: searches /refs/remotes/ and packed-refs.
GitUpdateProcess shows notification that branch doesn't exist yet and doesn't update.
GitPushActiveBranches correctly displays commit that are to be pushed.