GitNewChangesCollector: race condition in getting HEAD.
GitRepository is updated by the event of .git service files changes.
GitChangeProvider is requested for the changes independently of that.
Thus the GitChangeProvider may be requested for changes, before the GitRepository has the right value of the HEAD revision.
Thus force update the GitRepository to get the revision.

Comparing with native Git return value is removed - experiments show that the problem occurs.