GitUntrackedFilesHolder optimization
Don't ask Git for the status of possibly untracked files in a separate thread.
Instead keep track of "definitely untracked files" (which are added/removed to/from git index via IDEA actions) and "possibly untracked files" (e.g. when a file is created on disk and possibly not added to the index yet).
Check the status of suspicious files when GitNewChangesCollector asks for their status - thus using the thread of the GitChangeProvider.
The full refresh is initiated by GitNewChangesCollector as well - the need of the refresh is tracked by myReady, which is reset on startup and on .git/index change.