Fixed ClassLoader problem in GitConfig, patched ini4j
Section#as() called Proxy.newInstance with Thread#getContextClassLoader(), i.e. UrlClassLoader. To load an interface from git4idea a PluginClassLoader is needed.
Patch ini4j so that OptionMap#as() accepts ClassLoader. BeanTool.proxy uses passed ClassLoader or uses current in the case of null. Patch is attached to the sources.