Enable Git Integration: don't show dialog, just add roots & notify
Decided not to show the dialog proposing to create Git repository for the whole project,
because the proposal to create Git repository may be confusing.
Moreover, in this case subrepositories should be added to .gitignore of the super-repository,
and we are not going to do it by default.
A user who really wants such a complex configuration will have no problems with creating Git repository and adding the VCS root by hands.