Better fix for IDEA-78716 Master password prompt is "hidden"
Don't call invoke and wait in AWT, just run in that case.
Introduced GuiUtils.invokeAndWaitIfNeeded for this purpose, since we don't have a fair method in utils which invokes Application#invokeAndWait.

It still leads to an assertion error in Settings | GitHub, the fix for the assertion will follow in the next commit.