IDEA-89376 Git integration doesn't show change diff for moved file
Moved file has other path in the parent revision => constructing a fake GitFileRevision using the same file path is incorrect approach.
Modified the DiffFromHistoryHandler to pass the previous revision from the file history panel (and renamed the interface methods to avoid confusion between them). This is somewhat incorrect, because the previous revision in the list is not necessarily a direct parent of current revision (it is an ancestor though; the file can be not changed in the parent revision).
At best, we should query Git (git show <current revision>) to know the previous part, but it would require one more Git command. Current approach is acceptable until file history gets the graph log.

Merge commits, however, should be handled separately. The fix will come in the next commit.