[diff] Draw highlighters above gutters. Step 4: fix polygon colors for inline insertions/deletions
PROBLEM:
An inline insertion was shown as modification in the polygon, because the polygons are generated based on the type of the total fragment: a LineFragment, which contained children InlineFragments.

SOLUTION:
After inline child fragments are identified, adjust the type of their parent fragment if needed.
That is, if the change is just insertion or just deletion, the parent fragment should be of the same type. However, if the parent fragment contains two child changes of different types (e.g. an insertion and a deletion inside a single line), the summary change is a modification.