IDEA-92422 Unable to Annotate plain text files
When the UI Designer plugin is disabled, it doesn't register its own FileType, thus StdFileTypes.GUI_DESIGNER_FORM is precalculated as plain text.

Anyway, here is no reason anymore to prohibit Annotate on UI Designer forms, because there is a tab with XML, and annotate is perfectly valid for the XML.

This change undoes the change made for IDEA-34646.