Fix LineMarkersRenderer painting code in EditorGutterComponentImpl.
The rectangle height was calculated incorrectly, because endY was a top edge of the last line of the highlighted area, while it should be the bottom edge of it.
Therefore 1 line height should be added to the height.
However, if the highlighted text range ends with the end of the last line (including \n), the next line should not be included => don't add 1 line height for this case.

This also fixes IDEA-67658 & relates to IDEA-88946