[git] Add GitRepository parameter to the GitRepositoryChangeListener
GitRepositoryChangeListener is connected by a cross-repository Topic, thus adding it to one repository would actually add it to all of them.
Because of that and because of the possible need of removing a listener, get rid of addListener() methods (make clients connect by Topic instead), and add GitRepository as a listener method parameter.