WebServerManager: mark the server as started only if not disabled
Previously, if 1st unit test didn't set myEnabledInUnitTestMode to true,
but did request the WebServerManager somehow, started was set to true,
and the subsequent invocation wouldn't start the server, even if
myEnabledInUnitTestMode would be set to true by the next test.

Update the started variable only when server is actually set to start.