GitHub base test framework & the DialogManager
Introduce the DialogManager to be used instead of DialogWrapper#show
in the testable code.
Update & modify the TestDialogManager & TestDialogHandler.

Prepare a stub test for GithubShareAction.

Move github tests to a separate module with an extra dependency to
jira-connector - this is a temporary decision against a CNDFE
happening in tests. It will be removed when the problem is solved.