Fix github-tests project setup
github-tests were not running on the server, because nobody depended
on the module. The solution was either to write the module explicitly
to the main-idea-tests configuration, or to move the tests to github
module.

Tests were not moved to github module initially because they depend
on jira-connector (because tasks-tests depend on jira-connector).
Adding tasks-tests test dependency is impossible,
because they depend on testframework-java (which leads to prefix
problems).

Therefore it is decided to move github tests back to the github
module, and add jira-connector as a test dependency to it.