[vcs log] Async loading mini-details and details
If the value is not in the cache, load commit details around it in
the background.
Return an instance of LoadingDetails immediately,
and put it into the cache.
The calling code (the table, the changes browser) displays some
loading progress if receives LoadingDetails instead of a real commit.
It also subscribes to loading-finishes events to be able to refresh
the view once commit details have been loaded.

Loading happens on a QueueProcessor's pooled thread.
New tasks are added to the head of the queue, because it is more
actual to load details for the most recently requested commit.
Don't queue more than 10 loading processes: most likely the user has
just scrolled away from it, and he doesn't need these values right now.

Simplify VcsCommitCache: use it only from the DataGetter,
only via the described algorythm => no need to create a value if it
is not in the cache: DataGetter does it better. Also make sure that
any work with the cache is made from the EDT to avoid threading issues.

Introduce VcsLogProvider#readMiniDetails not to store big GitCommit
objects in the mini-details-cache (performance is now the same for
full- and mini-details).