[log] Improve synchronization & data flow in the VcsLogDataHolder
Encapsulate all information about the log data into a separate LogData.
VcsLogDataHolder is thread-safe, since all write-operations are
performed from a single sequential background queue,
and myLogData is substituted by new object on every change.

Previously the LogData it contained only the full log,
now it contains either full log, or the first part of it until the
full log is loaded; the DataPack built on this log; later it will
contain more.

Temporary remove the code which tried to guess the number of
top-commits which should be passed to DataPack.build: it is
incorrect, we have to remember this number per root. This will be
fixed in the next commit.