[log] Introduce filtering
I. Summary

We are able to filter graph by branches: not-wanted branches will be
just hidden with all their commits (unless they belong to a branch
that is to be shown).

We are not able to filter graph by other details yet: filtering, say,
by user can make "holes" in the graph, and we are not ready to
support this now.

Therefore, if user filters just by branch, we continue to show the
graph to him, but if he switches on a filter which is not supported
by graph (user, full-text), we turn the graph off by substituting the
TableModel with NoGraphTableModel which just shows commit subject in
the first column.

II. UI

Copy the same UI as it were in the old Git Log: a text field for
full-text search, popup menus for filtering by branch and by user.
Structure filter is not implemented yet.

The code of these classic-style filters seems to be tighly connected
with the old Log infrastructure => they are reimplemented from scratch.

Add a few niceties to popup controls: let them be focusable,
draw the focus border around them, make them keyboard friendly.

III. Details

VcsLogFilterer:
* VcsLogFilterer applies filters to the log. At first applies to the
  graph (to be able to filter by branch), then it applies other
  filters to details.
* The whole graph which was built can be filtered by branch. However
  details only of the recent commits are filtered by text/user, because
  details of older commits are not available.
  Get these details via VcsLogDataHolder#getTopCommitDetails().
* TODO: If there will be no or little results, we should load more
  commit details from the VCS. This will be implemented later.

VcsLogUI:
* Don't store the table model in VcsLogUI: it can be applied to the
  table at once.
* To keep filtering after refresh, apply filters instead of just
  creating new model.