[log] IDEA-115481 Fix permanent "loading" state
Count each task for loading more details sequentially, when creating it.

If then we get a LoadingDetails object from the cache,
which is very old, it means that its task was skipped
because of our limitation to avoid execution of all loading tasks (the
limitation is valid, because we don't want to have hundreds of tasks
for commits which user just scrolled through).

In that case remove the fake LoadingDetails object from the cache,
and let a new task to be submitted.