[log] Don't allow the graph table to repaint during graph modification
Introduce executeWithoutRepaint() which freezes the repaint while
executing the given task.

Otherwise it may lead to a performance disaster: the graph will
change, and current visible area will be not in the invalid table area,
which will lead to the full table repaint (which in turn will lead to
call getValueAt on all table rows).