[log] Compact Changes in VcsFullCommitDetails
Change is an expensive object, in particular because it extensively
uses FilePathImpl which always stores the path.

Use the compacted version of the Change in VcsFullCommitDetailsImpl:
LightChange can store only one VirtualFile or path (or two in case of
move);
Convert back to the Change when getChanges is called. All necessary
information (this and parent revisions, etc.) can be obtained from the
VcsFullCommitDetails object.

Unfortunately GitContentRevision and HgContentRevision require more
data for creation, therefore ContentRevisionFactory was introduced,
and passed to createFullCommitDetails factory method.
It handles creation of the VCS-plugin-specific ContentRevision object.