[log] Hide the Branches Panel by default, introduce log quick settings
Motivation:
The panel with branch labels and ability to navigate to the selected
branch by clicking on it seems to be not designed well enough,
especially when there are many local branches, and in the multi-repo
case.

Introduce "Quick Settings" button at the log toolbar.
Add "Show Branches Panel" action there.