[log] IDEA-116399 Fix filter by branch
--branches parameter doesn't work well: the pattern syntax it
supports is rather weak, and won't allow to effectively filter by
several branches. It also doesn't support filtering by a single
branch without blog parameters.

Therefore, since we allow to filter only by a single branch in the
log UI, assume it in the GitLogProvider, and use the <branch>
parameter (`git log origin/master`), which works fine.

If we will provide ability to filter by several branches someday (by
AND or by OR), we could use the graph capabilities to implement it.