[git log] Fix multi-value filter by user
Escaping baskslashes in 7eea76a introduced the problem with "\|"
separator used by Git to allow several users.

Backslashes should be escaped separately for text filter
and user filter: in the latter they should be escaped as well
(to allow a non-likely but possible case with backslash in user name)
before escaping "|".