[git] amend: don't add \n to the commit message
Unfortunately, git log returns the commit message always with \n at
the end.
Therefore there is no way to identify whether it was user who put \n
or it is added by Git.

The workaround is to not substitute commit message,
if it differs only with a trailing \n. It should suite most cases.