IDEA-120993 diff: fix diff range calculation in case of newline at the beginning of file
In case of empty range of one side we must return TextRange(n, n-1).

Previously, it was made by correctRangeEnd():
begin/end offsets are at the start of the same line -> we shift end one char backwards (because previous symbol is "\n") -> we got what we want
But in case of first line there are no "\n" that will move us from 0 line to -1 (because this is beginning of file).