[log] IDEA-117553 optimize git log initialization & refresh
Don't walk down from tags.
Passing --tags to git log significantly slows down the command, if
there are many tags in the repository (which is common).
On the other hand, it makes sense only if there are heads without branch
or HEAD labels (which is rare).

On initialization: don't consider tags. It may cause the log to
blick, when the full log will be loaded and built and if tagged
branches existed, were in the first 1000 and their artifacts (nodes,
edges) were visible on the screen. We can survive it.

On refresh: check if new tags were created, check if they point to
commits which we didn't receive in the first 1000. If yes, then it is
such tagged-branches. Just query some commits walking down from these
tags and put them into the set of commits for refresh.

We need to pass previous & current refs for this algorythm, but we
don't want to expose it to all VcsLogProviders, since might be not
clear for the implementors why do they get these refs.
=> introduce a Requirements extension which is really passed to
VcsLogProviders.