Introduced PyCallSiteExpression for distinguishing a[b] and a[b](c) in the type checker (PY-13051)
PyTypeChecker.analyzeCallSite() was used for analyzing the return type
of the call in the same way for different expressions: a[b] and a[b](c)
as analyzeCallSite(a[b]) making it impossible to distinguish between the
two.

Now we pass a PyCallExpression for analyzing the function call instead
of a PyQualifiedExpression.